connection: make "struct connection" opaque
Users of this library shouldn't access most structs directly.  This is
too dangerous when you add or change features, since the whole struct
and its layout would be part of the libmpdclient ABI.

Make the struct opaque with a forward declaration in mpd/connection.h,
and provide its real declaration in internal.h.